feat: Create the config DB and implement logging.
This sets up the connection to the SQLite DB to be used for the config.
For now, the path to the config is in SSNFSd.pro.
Log paths are stored in the DB and each log can be configured to only get
certain categories of messages.
So, for each log one can enable or disable categories and individual
log levels per category.

Issues #10 and #11